Uniting auxiliary test structures into one
