+++ demos/cpp/demo-rtsolver-shell-test.cpp
@@ -19,10 +19,10 @@
 using namespace Reaktoro;
 
 // Reactive transport test includes
-#include <demos/cpp/RTSolverTest.h>
+#include <demos/cpp/TestUtils.h>
 
 /// Forward declaration
-auto runReactiveTransport(Params& params, Results& results) -> void;
+auto runReactiveTransport(ReactiveTransportParams& params, ReactiveTransportResults& results) -> void;
 
 int main()
 {
@@ -32,7 +32,7 @@ int main()
     int minute = 60;
 
     // Step 2: Define parameters for the reactive transport simulation
-    Params params;
+    ReactiveTransportParams params;
 
     // Define discretization parameters
     params.xl = 0.0; // the x-coordinates of the left boundaries
@@ -76,7 +76,7 @@ int main()
     params.outputConsole();
 
     // Results
-    Results results;
+    ReactiveTransportResults results;
 
     // Execute reactive transport with different solvers
     params.use_smart_equilibrium_solver = true; runReactiveTransport(params, results);
@@ -112,7 +112,7 @@ int main()
 
     return 0;
 }
-auto runReactiveTransport(Params& params, Results& results) -> void
+auto runReactiveTransport(ReactiveTransportParams& params, ReactiveTransportResults& results) -> void
 {
     // Step **: Create the results folder
     auto folder = params.makeResultsFolder("shell-test");
